+++ sunpy/coordinates/tests/test_frames.py
@@ -10,15 +10,14 @@ from astropy.tests.helper import assert_quantity_allclose
 
 from astropy.coordinates import (UnitSphericalRepresentation,
                                  SphericalRepresentation,
-                                 CartesianRepresentation)
+                                 CartesianRepresentation,
+                                 SkyCoord)
 
 from ... import sun
 from ..frames import (Helioprojective,
                       HeliographicStonyhurst,
                       Heliocentric,
                       HeliographicCarrington)
-from ..representation import (UnitSphericalWrap180Representation,
-                              SphericalWrap180Representation)
 
 RSUN_METERS = sun.constants.get('radius').si.to(u.m)
 DSUN_METERS = sun.constants.get('mean distance').si.to(u.m)
@@ -39,11 +38,10 @@ These are common 2D params, kwargs are frame specific
 two_D_parameters = [
     ([0 * u.deg, 0 * u.arcsec], None),
     ([0 * u.deg, 0 * u.arcsec], {'obstime': '2011/01/01T00:00:00'}),
-    ([0 * u.deg, 0 * u.arcsec], {'representation': 'unitsphericalwrap180'}),
     ([0 * u.deg, 0 * u.arcsec], {'representation': 'unitspherical'}),
-    ([UnitSphericalWrap180Representation(0 * u.deg, 0 * u.arcsec)], None),
+    ([UnitSphericalRepresentation(0 * u.deg, 0 * u.arcsec)], None),
     ([UnitSphericalRepresentation(0 * u.deg, 0 * u.arcsec)], None), (
-        [UnitSphericalWrap180Representation(0 * u.deg, 0 * u.arcsec)],
+        [UnitSphericalRepresentation(0 * u.deg, 0 * u.arcsec)],
         {'obstime': '2011/01/01T00:00:00'})
 ]
 """
@@ -52,13 +50,11 @@ These are common 3D params, kwargs are frame specific
 three_D_parameters = [
     ([0 * u.deg, 0 * u.arcsec, 1 * u.Mm], None),
     ([0 * u.deg, 0 * u.arcsec, 1 * u.Mm], {'obstime': '2011/01/01T00:00:00'}),
-    ([0 * u.deg, 0 * u.arcsec, 1 * u.Mm], {'representation': 'sphericalwrap180'
-                                           }),
     ([0 * u.deg, 0 * u.arcsec, 1 * u.Mm], {'representation': 'spherical'}),
-    ([SphericalWrap180Representation(0 * u.deg, 0 * u.arcsec, 1 * u.Mm)],
+    ([SphericalRepresentation(0 * u.deg, 0 * u.arcsec, 1 * u.Mm)],
      None),
     ([SphericalRepresentation(0 * u.deg, 0 * u.arcsec, 1 * u.Mm)], None), (
-        [SphericalWrap180Representation(0 * u.deg, 0 * u.arcsec, 1 * u.Mm)],
+        [SphericalRepresentation(0 * u.deg, 0 * u.arcsec, 1 * u.Mm)],
         {'obstime': '2011/01/01T00:00:00'})
 ]
 
@@ -82,7 +78,7 @@ def test_create_hpc_2d(args, kwargs):
         assert isinstance(hpc1._data, UnitSphericalRepresentation)
     else:
         # Check that we have a 2D wrap180 representation
-        assert isinstance(hpc1._data, UnitSphericalWrap180Representation)
+        assert isinstance(hpc1._data, UnitSphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc1.Tx == 0 * u.arcsec
@@ -111,7 +107,7 @@ def test_create_3d(args, kwargs):
         assert isinstance(hpc1._data, SphericalRepresentation)
     else:
         # Check that we have a 2D wrap180 representation
-        assert isinstance(hpc1._data, SphericalWrap180Representation)
+        assert isinstance(hpc1._data, SphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc1.Tx == 0 * u.arcsec
@@ -121,7 +117,7 @@ def test_create_3d(args, kwargs):
     # Check the attrs are in the correct default units
     assert hpc1.Tx.unit is u.arcsec
     assert hpc1.Ty.unit is u.arcsec
-    assert hpc1.distance.unit is u.km
+    assert hpc1.distance.unit is u.Mm
 
 
 def test_cart_init():
@@ -132,46 +128,6 @@ def test_cart_init():
     assert isinstance(hpc1._data, CartesianRepresentation)
 
 
-# # This is not actually valid, it should be re-purposed to be Heliocentric
-# # cylindrical Assuming that that is a vaild representation.
-# cylindrical_parameters = [
-#     ([100 * u.km, 25 * u.deg, 1 * u.Mm], {'representation': 'cylindrical'}), (
-#         [100 * u.km, 25 * u.deg, 1 * u.Mm], {'obstime': '2011/01/01T00:00:00',
-#                                              'representation': 'cylindrical'}),
-#     ([100 * u.km, 25 * u.deg], {'distance': 1 * u.Mm,
-#                                 'representation': 'cylindrical'}),
-#     (None, {'rho': 100 * u.km,
-#             'psi': 25 * u.deg,
-#             'distance': 1 * u.Mm,
-#             'representation': 'cylindrical'}),
-#     ([CylindricalRepresentation(100 * u.km, 25 * u.deg, 1 * u.Mm)],
-#      {'representation': 'cylindrical'}), (
-#          [CylindricalRepresentation(100 * u.km, 25 * u.deg, 1 * u.Mm)],
-#          {'obstime': '2011/01/01T00:00:00',
-#           'representation': 'cylindrical'})
-# ]
-#
-#
-# @pytest.mark.parametrize('args, kwargs', cylindrical_parameters)
-# def test_create_cylindrical(args, kwargs):
-#     hpc1 = init_frame(Helioprojective, args, kwargs)
-#
-#     # Check we have the right class!
-#     assert isinstance(hpc1, Helioprojective)
-#     # Check that we have a 2D wrap180 representation
-#     assert isinstance(hpc1._data, CylindricalRepresentation)
-#
-#     # Check the attrs are correct
-#     assert hpc1.rho == 100 * u.km
-#     assert hpc1.psi == 25 * u.deg
-#     assert hpc1.distance == 1 * u.Mm
-#
-#     # Check the attrs are in the correct default units
-#     assert hpc1.rho.unit is u.km
-#     assert hpc1.psi.unit is u.arcsec
-#     assert hpc1.distance.unit is u.km
-
-
 # Test HPC Calculate Distance
 def test_hpc_distance():
     hpc1 = Helioprojective(0 * u.deg, 0 * u.arcsec,
@@ -179,7 +135,7 @@ def test_hpc_distance():
 
     assert isinstance(hpc1, Helioprojective)
     # Check that we have a 2D wrap180 representation
-    assert isinstance(hpc1._data, UnitSphericalWrap180Representation)
+    assert isinstance(hpc1._data, UnitSphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc1.Tx == 0 * u.arcsec
@@ -187,7 +143,7 @@ def test_hpc_distance():
 
     hpc2 = hpc1.calculate_distance()
 
-    assert isinstance(hpc2._data, SphericalWrap180Representation)
+    assert isinstance(hpc2._data, SphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc2.Tx == 0 * u.arcsec
@@ -201,7 +157,7 @@ def test_hpc_distance_off_limb():
 
     assert isinstance(hpc1, Helioprojective)
     # Check that we have a 2D wrap180 representation
-    assert isinstance(hpc1._data, UnitSphericalWrap180Representation)
+    assert isinstance(hpc1._data, UnitSphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc1.Tx == 1500 * u.arcsec
@@ -209,7 +165,7 @@ def test_hpc_distance_off_limb():
 
     hpc2 = hpc1.calculate_distance()
 
-    assert isinstance(hpc2._data, SphericalWrap180Representation)
+    assert isinstance(hpc2._data, SphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc2.Tx == 1500 * u.arcsec
@@ -222,7 +178,7 @@ def test_hpc_distance_3D():
 
     assert isinstance(hpc1, Helioprojective)
     # Check that we have a 2D wrap180 representation
-    assert isinstance(hpc1._data, SphericalWrap180Representation)
+    assert isinstance(hpc1._data, SphericalRepresentation)
 
     # Check the attrs are correct
     assert hpc1.Tx == 1500 * u.arcsec
@@ -231,6 +187,20 @@ def test_hpc_distance_3D():
     hpc2 = hpc1.calculate_distance()
 
     assert hpc2 is hpc1
+
+
+def test_wrapping_on():
+    hpc1 = Helioprojective(359.9*u.deg, 10*u.deg)
+    assert_quantity_allclose(hpc1.Tx, -0.1*u.deg)
+    assert_quantity_allclose(hpc1.Tx.wrap_angle, 180*u.deg)
+
+
+def test_wrapping_off():
+    hpc1 = Helioprojective(359.9*u.deg, 10*u.deg, wrap_longitude=False)
+    assert_quantity_allclose(hpc1.Tx, 359.9*u.deg)
+    assert_quantity_allclose(hpc1.Tx.wrap_angle, 360*u.deg)
+
+
 
 # ==============================================================================
 # ## Heliographic Tests
@@ -252,7 +222,7 @@ def test_create_hgs_2d(frame, args, kwargs):
         assert isinstance(hgs1._data, SphericalRepresentation)
     elif isinstance(hgs1, HeliographicStonyhurst):
         # Check that we have a 2D wrap180 representation
-        assert isinstance(hgs1._data, SphericalWrap180Representation)
+        assert isinstance(hgs1._data, SphericalRepresentation)
 
     # Check the attrs are correct
     assert hgs1.lon == 0 * u.deg
@@ -275,9 +245,7 @@ def test_create_hgs_force_2d(frame, args, kwargs):
 
     rep_kwarg = kwargs.get('representation', None) if kwargs else None
 
-    if rep_kwarg == 'unitsphericalwrap180':
-        assert isinstance(hgs1._data, UnitSphericalWrap180Representation)
-    elif rep_kwarg == 'unitspherical':
+    if rep_kwarg == 'unitspherical':
         assert isinstance(hgs1._data, UnitSphericalRepresentation)
 
     # Check the attrs are correct
@@ -314,7 +282,7 @@ def test_create_hgs_3d(frame, args, kwargs):
             assert isinstance(hgs1._data, SphericalRepresentation)
         elif isinstance(hgs1, HeliographicStonyhurst):
             # Check that we have a 2D wrap180 representation
-            assert isinstance(hgs1._data, SphericalWrap180Representation)
+            assert isinstance(hgs1._data, SphericalRepresentation)
 
     # Check the attrs are correct
     assert hgs1.lon == 0 * u.deg
@@ -334,6 +302,23 @@ def test_hgs_cart_init():
 
     assert isinstance(hpc1, HeliographicStonyhurst)
     assert isinstance(hpc1._data, CartesianRepresentation)
+
+
+def test_hgs_wrapping_on():
+    hpc1 = HeliographicStonyhurst(350*u.deg, 10*u.deg)
+    assert_quantity_allclose(hpc1.lon, -10*u.deg)
+    assert_quantity_allclose(hpc1.lon.wrap_angle, 180*u.deg)
+
+
+def test_hgs_wrapping_off():
+    hpc1 = HeliographicStonyhurst(350*u.deg, 10*u.deg, wrap_longitude=False)
+    assert_quantity_allclose(hpc1.lon, 350*u.deg)
+    assert_quantity_allclose(hpc1.lon.wrap_angle, 360*u.deg)
+
+def test_hgc_wrapping_360():
+    hpc1 = HeliographicCarrington(350*u.deg, 10*u.deg)
+    assert_quantity_allclose(hpc1.lon, 350*u.deg)
+    assert_quantity_allclose(hpc1.lon.wrap_angle, 360*u.deg)
 
 
 # ==============================================================================
@@ -362,3 +347,58 @@ def test_create_hcc_3d(args, kwargs):
     assert hcc.x.unit is u.km
     assert hcc.y.unit is u.km
     assert hcc.z.unit is u.km
+
+# ==============================================================================
+# SkyCoord Tests
+# ==============================================================================
+
+
+two_D_parameters = [
+    ([0 * u.deg, 0 * u.arcsec], {}),
+    ([UnitSphericalRepresentation(0 * u.deg, 0 * u.arcsec)], {}),
+    ([UnitSphericalRepresentation(0 * u.deg, 0 * u.arcsec)], {}),
+    ([SphericalRepresentation(0 * u.deg, 0 * u.arcsec, 1*u.one)], {}),
+]
+
+
+@pytest.mark.parametrize("args, kwargs",
+                         two_D_parameters + [([0 * u.deg, 0 * u.arcsec],
+                                              {'representation': 'unitspherical'})])
+def test_skycoord_hpc(args, kwargs):
+    """
+    Test that when instantiating a HPC frame with SkyCoord calculate distance
+    still works.
+    """
+
+    # Python 3: These should just be keywords in the `SkyCoord` constructor
+    kwargs.update({
+        "frame": "helioprojective",
+        "obstime": "2011-01-01T00:00:00"
+        })
+    sc = SkyCoord(*args, **kwargs)
+    # Test the transform to HGS because it will force a `calculate_distance` call.
+    hgs = sc.transform_to("heliographic_stonyhurst")
+
+    assert isinstance(hgs.frame, HeliographicStonyhurst)
+
+
+@pytest.mark.parametrize("args, kwargs", two_D_parameters)
+def test_skycoord_hgs(args, kwargs):
+    """
+    Test that when instantiating a HPC frame with SkyCoord correctly replaces
+    distance.
+
+    Note: We only need to test HGS here not HGC as they share the same
+    constructor.
+    """
+
+    RSUN_METERS = sun.constants.get('radius').si
+
+    # Python 3: These should just be keywords in the `SkyCoord` constructor
+    kwargs.update({
+        "frame": "heliographic_stonyhurst",
+        "obstime": "2011-01-01T00:00:00"
+        })
+    sc = SkyCoord(*args, **kwargs)
+
+    assert_quantity_allclose(sc.radius, RSUN_METERS)