+++ CHANGELOG.rst
@@ -4,9 +4,24 @@
 New Features
 ------------
 
+- Added TimeUTime class to support utime. [#2409]
+
+API Changes
+-----------
+
+- ``sunpy.coordinates.representation`` has been removed. Longitude wrapping is
+  now done in the constructor of the frames. [#2431]
+- Propagation of ``obstime`` in the coordinate frame transformation has changed,
+  this means in general when transforming directly between frames (not
+  ``SkyCoord``) you will have to specify ``obstime`` in more places. [#2461]
+- Transforming between Heliographic Stonyhurst and Carrington now requires that
+  ``obstime`` be defined and the same on both the input and output frames. [#2461]
+
 Bug Fixes
 ---------
 
+- Fix HGS frame constructor and HPC ``calculate_distance`` with SkyCoord constructor. [#2463]
+- removed `wavelnth` keyword in meta desc of Maps to avoid using non standard FITS keyword like `nan` [#2427]
 - `~sunpy.net.dataretriever.clients.XRSClient` now reports time ranges of files correctly. [#2364]
 - Make parse_time work with datetime64s and pandas series [#2370]
 - CompositeMap axes scaling now uses map spatial units [#2310]